Removing docker-compose in favor of Make
`docker-compose` is not a good tool for building images. Its primary
use is to deploy containers. Most advice online suggests using Make to
automate the building of images.

Those wishing to build the image should run:

```
make
```

The image may be pushed to our repository (assuming correct permissions)
with:

```
make push
```

The Dockerfiles have been refactored to always compile and build that on
the stable branch. This removes the necessity of updating versions here
as we release newer versions of gem5.